requirements.txt
Flask
Flask-Cors
Flask-SocketIO
Flask-SQLAlchemy
psycopg2
requests
python-dotenv